Optimal weighting Spreadsheet
Mathematically optimize portfolio wealth allocation, all the math is done.  It works with any infinite number of assets, and will tell us the percentage of portfolio wealth to put into each bet.  i use the excel>data tab>solver(add-in) to have it 'solve' and fill in the values so we have our money best allocated.